Moved SliceData to its own file.
* `data_write.SliceData` -> `utils.file_formats.SliceData`
* Refactored the general HDF5 file format:
   - Top level "metadata" group
   - All data stored as `Dataset`, with `"description"` metadata attached
   - File-level metadata is hard-linked within each record group to the top-level "metadata" group entry
